Refactor tests into separate classes and interfaces
