--- test/src/macosTest/kotlin/io/realm/shared/NotificationTests.kt
@@ -1,1 +1,0 @@
-../../../../../androidTest/kotlin/io/realm/shared/NotificationTests.kt
