+++ test/src/macosTest/kotlin/io/realm/shared/notifications
@@ -1,0 +1,1 @@
+../../../../../androidTest/kotlin/io/realm/shared/notifications
